TextLayout:  Fix UTF-16 Text in Cppia
Resolves issues seen here: 
https://github.com/openfl/openfl/issues/2264

Partially addresses issues here:
https://github.com/openfl/lime/pull/1916
https://github.com/openfl/lime/pull/1917

This should work on cpp and cppia though, we will continue to use untyped cpp for now. We should provide a fix for this in the native code since it will avoid expensive O(n) conversions to force this to work. For now, this provides a necessary hotfix.

Passing a raw String directly to hb_buffer_add_utf16() caused issues because cpp/cppia's internal memory layout does not match what HarfBuzz expects. To address this, we utilize UInt16Array to directly stores charCodeAt(i) as native 16-bit values. This means HarfBuzz gets properly-aligned UTF-16 input.

On cpp, we used something like untyped __cpp__("text.wc_str()"), but cppia lacks the inherent ability to execute untyped cpp. The new approach mimics native UTF-16 strings perfectly, making it work without needing native C++ access.

Once again, this isn't an ideal fix, since it requires O(n) traversal of strings and additional allocations, but it does make sure we get the appropriate behavior.

We still need to address UTF-8 issues on hxcpp in lime in the PR's referenced earlier, but with this, we no longer need to rely on UTF-8 for cppia.