TextField: preserve selection on mouse up after mouse down and drag
Keep existing __lineSelection and __wordSelection behavior because I'm not sure
if there are edge cases that might break.

However, for regular mouse down a drag, the existing selection might have been
provided by setSelection(), which we don't want to change unless the mouse
actually moves, or the scroll position has changed (which is like the mouse
moving).