Object: on JS target, use Function.prototype.bind() when returning a function with a.b member access
We don't want the value of `this` to get lost when calling a.b() as a method.

Example of Haxe code that fails with JS target:

```haxe
function(a:openfl.utils.Object):Void
{
	a.b();
}
```

Interestingly, when assigning to Dynamic first, the call starts working as expected.

```haxe
function(a:openfl.utils.Object):Void
{
	var dyn:Dynamic = a;
	dyn.b();
}
```

hxcpp doesn't seem to have the same issue as the JS target, and both examples above work.

Normally, in JS runtimes, calling `a.b()` assigns `a` to `this` in the context of the called method. However, the generated JS involving `openfl.utils.Object.__get()` ends up being equivalent to this instead:

var b = a.b;
b();

The JS runtime won't assign anything to `this` in the case where `b()` is called without member access, so `this` is undefined. That can be unexpected when `Dynamic` works and `openfl.utils.Object` is supposed to be at least somewhat similar to `Dynamic`.

We don't have control over the JS generated by the Haxe compiler here. However, we can use the JS function Function.prototype.bind() to keep the method's expected `this` value.

It's also worth mentioning that Flash/AS3 keeps the `this` value when assigning a method to a variable, like `var b:Function = a.b;` and then calling `b()` without member access, which is notably different from JS. However, since the purpose of `openfl.utils.Object` is to keep some of AS3's common coding practices working with OpenFL, this use of `bind()` actually fits in very well as functionality of `openfl.utils.Object`.