scale9Grid: New implementation that matches Flash's behavior much more closely
When drawing with Graphics, the position of each point is now adjusted based on where it falls within the scale9Grid. Previously in OpenFL, Graphics was basically drawn as a bitmap and sliced up, which could lead to a single gradient or bitmap fill that spans multiple regions of the scale9Grid getting scaled inconsistently. Now, when using scale9Grid, the entire fill is scaled the same throughout when it spans multiple regions. The scale of each fill is basically calculated based on how much the distance between the min and max points of the current shape that it is drawn within has changed.

The previous implementation of scale9Grid remains available for now, but is disabled by default. It may be enabled with a new openfl_legacy_scale9grid define. This legacy implementation will likely be removed in a future version of OpenFL.