Context3DGraphics: fix when drawTriangles() should be skipped on hardware
Two cases.

If using beginFill(), uvtData must always be null. If it is non-null, SWF, Canvas, and Cairo skip it entirely.

Currently, OpenFL can't draw bitmaps/shaders in hardware when uvtData is null. If not forcing hardware, we delegate to software. However, when forcing hardware, we should skip it, like we skip other things when they are not yet supported in hardware. This case can be considered temporary. If someone ever implements bitmap/shader fills for drawRect(), we should be able to support them with drawTriangles() too.