Context3DGraphics: implement drawRect() with bitmap or shader fill
Uses the drawTriangles() implementation. Other shapes could potentially use drawTriangles() in the future too, like rounded rectangles, ellipses, or circles, if we can triangulate them.

Left the solid color implementation of drawRect() the same to avoid breaking anything, but it actually could be done using the same drawTriangles() implementation too. We might switch later if it turns out to be faster and doesn't break anything.