CanvasGraphics, CanvasTextField: optimize by re-using the existing BitmapData, if the width and height has not changed
The new BitmapData would be using the same underlying canvas, so it makes sense to keep using it, if we can.

This helps to avoid frequent garbage collection of BitmapData, Image, and GPU textures.

In a benchmark that cleared and redrew graphics at the same dimensions every frame, I saw a measurable increase in FPS on some older devices.