TextField: on html5, after creating the Graphics objects, clear its command buffer
OpenFL adds an implicit moveTo(0, 0) to Graphics on the html5 target because the
HTML Canvas API acts a bit quirky if it is missing. See #531 for details.

However, in TextField, we never add any commands to Graphics. We just work with
the underlying canvas. So we don't want to try to process the command
buffer at all.

This didn't actually cause any issues in the past (except perhaps affecting
performance by calling functions that didn't need to be called). However,
starting in commit 65abbfe1977a165daa3329f564e27ef34b31b8cc, text started to be
rendered at the wrong pixel ratio on html5 only. That's when I discovered that
it was trying to process the graphics command buffer that would always have a
single MOVE_TO command on html5, but no commands on native. The CanvasGraphics
class was changing the bitmap scale to 1 that the CanvasTextField class had
already set to pixelRatio.