CanvasGraphics, CanvasTextField: when replacing BitmapData, dispose old texture immediately (closes #2748)
When creating a new BitmapData with the same canvas, we can dispose the existing texture because it will no longer be used. This frees up memory immediately instead of waiting for garbage collection to do it automatically later.

Unlike #2748, we don't dispose the BitmapData because that will now clear/reset the underlying Canvas, with the merge of PR #2755. The same canvas from the first BitmapData is reused by the second BitmapData, so we can't let it get disposed.

Co-authored-by: barisyild <barisyild@users.noreply.github.com>