Context3DGraphics: Fix render() not setting graphics.__hardwareDirty to false when it takes the software path (closes #2762)
The flag should be cleared regardless of whether Context3DGraphics is drawing the graphics on its own or delegating to Canvas/Cairo and displaying that as a texture.

Otherwise, the __hardwareDirty flag will remain set indefinitely, which can cause an unnecessary re-render every frame.

To explain a different way, updating the texture is what needs to be done when hardware is dirty. Once that's done, hardware is no longer dirty.

Additionally, it's worth mentioning that Context3DTextField.render() always draws in software (there is no hardware path for text in OpenFL at this time), but it sets graphics.__hardwareDirty to false regardless.