BitmapData: don't dispose the canvas and context on html5 when the BitmapData was created by the Assets class
Commit b0ee945d751ed99ac64f303f61165ced6799c180 (#2755) started disposing the canvas/context on html5 to fix a memory leak on Safari (and it actually frees up memory faster for all browsers, which is a better implementation that more closely matches how dispose() is documented to work). However, disposing BitmapData that was created by Assets.getBitmapData() unexpectedly resulted in new exceptions being thrown.

On most targets, Lime's Assets.getImage() returns a new Image every time, and modifying that Image has no affect on future calls to Assets.getImage(). On html5, it always returns the same Image, which means that modifying or disposing the Image's contents will affect future calls to Assets.getImage(). In particular, after clearing/resizing the canvas/context, future calls to Assets.getImage() will return an invalid Image on html5 because a fresh canvas/context isn't created every time.

So now we don't dispose the canvas/context specifically when the BitmapData was created by Assets.getBitmapData(). This is not really ideal, but it's probably better to maintain backwards compatibility in this case, at least for now. In particular, the change seems to affect Flixel, which is one of the bigger libraries that depends on OpenFL/Lime in our ecosystem.

I can't currently say which behavior that I would prefer, but making those methods more consistent is something we should consider for future major updates of Lime and OpenFL. The html5 target seems to be the trickiest. Technically, it can create new Images synchronously by cloning, as long as it has completed creating the first one asynchronously, but it's pretty expensive performance-wise. So making all targets return a single shared Image might be the better choice, if we want to make all targets more consistent. In that case, libraries like Flixel will need to remember which BitmapData instances come from Assets.getBitmapData(), and they'll need to avoid disposing the assets BitmapData when it might still be needed. I'm no Flixel expert, but it looks to me like FlxGraphic may be able add an internal flag to track whether the BitmapData came from from Assets or not. Similar to what we're using as a workaround in this commit.