actions: use macos-latest on unit-test-hashlink, use latest krdlab/setup-haxe on all jobs
It is using commit hash krdlab/setup-haxe@8f35d1215b93e940a76f9470e22e8a5ba6149598 for now because it hasn't be included in a release tag yet. However, macos-13 is getting removed in December, and this commit is necessary to use arm64 macOS with certain Haxe versions. I'd rather get off macos-13 sooner than later, and we can switch to a tag later when there is one.